+++ src/DonorProfiles/resources/js/app/components/auth-modal/utils/index.js
@@ -13,9 +13,10 @@ export const loginWithAPI = ( { login, password } ) => {
 		.then( ( response ) => response.data );
 };
 
-export const verifyEmailWithAPI = ( { email } ) => {
+export const verifyEmailWithAPI = ( { email, recaptcha } ) => {
 	return axios.post( getAPIRoot() + 'give-api/v2/donor-profile/verify-email', {
 		email,
+		'g-recaptcha-response': recaptcha,
 	}, 	{
 		headers: {
 			'X-WP-Nonce': getAPINonce(),